Move validation code into separate module
